+++ services/editor/package.json
@@ -84,6 +84,7 @@
     "@testing-library/react-hooks": "^5.1.2",
     "@types/adal-angular": "^1.0.1",
     "@types/jest": "^26.0.9",
+    "@types/md5": "^2.3.0",
     "@types/node": "^15.0.1",
     "@types/ramda": "0.27.40",
     "@types/react": "^17.0.4",
