+++ services/editor/src/components/UserBar.tsx
