+++ services/editor/src/contexts/Tweek.tsx
@@ -1,16 +1,17 @@
-import React, { useEffect, useState } from 'react';
-import { prepareKey, TweekProvider as OriginalTweekProvider } from 'react-tweek';
-import { TweekRepository } from 'tweek-local-cache';
 import { equals } from 'ramda';
+import React, { FunctionComponent, useEffect, useState } from 'react';
+import { prepareKey, TweekProvider as OriginalTweekProvider } from 'react-tweek';
+import { CurrentUser } from 'tweek-client';
+import { TweekRepository } from 'tweek-local-cache';
 import { tweekClient } from '../utils';
 import { useCurrentUser } from './CurrentUser';
 
 prepareKey('@tweek/editor/_');
 
-const toTweekContext = ({ User }) => ({ tweek_editor_user: User });
+const toTweekContext = ({ User }: CurrentUser) => ({ tweek_editor_user: User });
 
-export const TweekProvider = ({ children }) => {
-  const [tweekRepository, setTweekRepository] = useState();
+export const TweekProvider: FunctionComponent = ({ children }) => {
+  const [tweekRepository, setTweekRepository] = useState<TweekRepository>();
 
   const user = useCurrentUser();
